+++ NEWS.md
@@ -1,5 +1,7 @@
 # cpp11 (development version)
 
+* Fix crash related to unwind protect optimization (#244)
+
 # cpp11 0.4.0
 
 ## New Features